Add simple argument parser to control parameter k and error rate.

+++ src/io_cfg_type.hpp
@@ -8,7 +8,6 @@
 
 #include <cstring>
 #include <experimental/filesystem>
-//#include <fstream>
 #include <iostream>
 #include <regex>
 
@@ -26,53 +25,33 @@ namespace priset
 
 struct io_cfg_type
 {
-private:
-    // Directory that contains library, taxonomy, and taxid to accession files.
-    fs::path lib_dir;
-    // Taxid to accession map in csv format (set by PriSeT).
-    fs::path acc_file{};
-    // Sequence library file in fasta format (set by PriSeT).
-    fs::path fasta_file{};
-    // 1-based IDs and accession numbers.
-    fs::path id_file{};
-    // Taxonomy file in csv format (set by PriSeT).
-    fs::path tax_file{};
-    // Working base directory for storing FM indices, mappings and results.
-    fs::path work_dir;
-    // Working subdirectory for FM index (set by PriSeT).
-    fs::path index_dir;
-    // Working subdirectory for FM index mappings (set by PriSeT).
-    fs::path mapping_dir;
-    // Path to genmap binary (set by PriSeT).
-    fs::path genmap_bin;
-    // Library file extensions.
-    std::string ext_fasta = ".fasta";
-    std::string ext_acc = ".acc";
-    std::string ext_tax = ".tax";
-    std::string ext_id = ".id";
-    // Path to R shiny app template
-    fs::path app_template = "../PriSeT/src/app_template.R";
-    // R script for launching shiny app.
-    fs::path script_runner;
-    // Path to generated copy of R script to be run in terminal.
-    fs::path script_file;
-    // Path to store result tables to load in Shiny.
-    fs::path result_file;
-    // Path to primer info file (sequences and chemical attributes)
-    fs::path primer_info_file;
 
 public:
-    // Give full path, no ~/ or ../!
-    io_cfg_type(fs::path const & lib_dir_, fs::path const & work_dir_) :
-        lib_dir{fs::canonical(lib_dir_)},
-        work_dir{fs::canonical(work_dir_)},
-        index_dir{fs::canonical(work_dir_)},
-        mapping_dir{fs::canonical(work_dir_)},
-        genmap_bin{fs::current_path()}
+    // Default constructor.
+    io_cfg_type() = default;
+
+    // Default copy constructor.
+    io_cfg_type(io_cfg_type const &) = default;
+
+    // Default copy construction via assignment.
+    io_cfg_type & operator=(io_cfg_type const &) = default;
+
+    // Move constructor.
+    io_cfg_type(io_cfg_type && rhs) = default;
+
+    // Move assignment.
+    io_cfg_type & operator=(io_cfg_type && rhs) = default;
+
+    // Set library and working directory paths.
+    void assign(fs::path const & lib_dir_, fs::path const & work_dir_, bool const skip_idx_)
     {
-        genmap_bin /= "submodules/genmap/bin/genmap";
+        lib_dir = fs::canonical(lib_dir_);
+        work_dir = fs::canonical(work_dir_);
+        skip_idx = skip_idx_;
+        index_dir = work_dir;
+        mapping_dir = work_dir;
+        genmap_bin = fs::current_path() / "submodules/genmap/bin/genmap";
         // parse library directory and assign paths to the .acc, .fasta, and .tax files
-        std::cout << "lib_dir = " << lib_dir << std::endl;
 //        std::cout << "current path: " << fs::current_path() << std::endl;
 
         if (!fs::exists(lib_dir))
@@ -113,7 +92,7 @@ public:
 
         // set output directory for FM index, will be created by genmap
         index_dir /= fs::path("/index");
-        if (fs::exists(index_dir))
+        if (!skip_idx && fs::exists(index_dir))
         {
             char cmd[50];
             sprintf(cmd, "rm -r %s", index_dir.c_str());
@@ -136,6 +115,7 @@ public:
         result_file = result_path / "results.csv";
         primer_info_file = result_path / "primer_info.csv";
         script_file = get_work_dir() / "app" / "app.R";
+        std::cout << "STATUS\tSet R script file: " << script_file << std::endl;
         script_runner = get_work_dir() / "app" / "app_run.R";
 
         if (!fs::exists(script_runner))
@@ -146,6 +126,15 @@ public:
             ofs.close();
         }
     };
+
+    // Destructor.
+    ~io_cfg_type() = default;
+
+    // Return skip_idx flag.
+    bool get_skip_idx() const noexcept
+    {
+        return skip_idx;
+    }
 
     // Return accession file with absolute path as filesystem::path object.
     fs::path get_acc_file() const noexcept
@@ -220,6 +209,7 @@ public:
     // Return path to R script file to be run in terminal.
     fs::path get_script_file() const noexcept
     {
+        std::cout << "Enter get_script_file, " << script_file << std::endl;
         return script_file;
     }
 
@@ -239,6 +229,44 @@ public:
     {
         return work_dir;
     }
+
+private:
+    // Directory that contains library, taxonomy, and taxid to accession files.
+    fs::path lib_dir;
+    // Working base directory for storing FM indices, mappings and results.
+    fs::path work_dir;
+    // Flag for indicating if index computation shall be skipped (because it already exists).
+    bool skip_idx;
+
+    // Taxid to accession map in csv format (set by PriSeT).
+    fs::path acc_file{};
+    // Sequence library file in fasta format (set by PriSeT).
+    fs::path fasta_file{};
+    // 1-based IDs and accession numbers.
+    fs::path id_file{};
+    // Taxonomy file in csv format (set by PriSeT).
+    fs::path tax_file{};
+    // Working subdirectory for FM index (set by PriSeT).
+    fs::path index_dir;
+    // Working subdirectory for FM index mappings (set by PriSeT).
+    fs::path mapping_dir;
+    // Path to genmap binary (set by PriSeT).
+    fs::path genmap_bin;
+    // Library file extensions.
+    std::string ext_fasta = ".fasta";
+    std::string ext_acc = ".acc";
+    std::string ext_tax = ".tax";
+    std::string ext_id = ".id";
+    // Path to R shiny app template
+    fs::path app_template = "../PriSeT/src/app_template.R";
+    // R script for launching shiny app.
+    fs::path script_runner;
+    // Path to generated copy of R script to be run in terminal.
+    fs::path script_file;
+    // Path to store result tables to load in Shiny.
+    fs::path result_file;
+    // Path to primer info file (sequences and chemical attributes)
+    fs::path primer_info_file;
 
 };
